CRUD is complete with auth and crypt

+++ backend/controllers/goalController.js
@@ -2,11 +2,12 @@ const asyncHandler = require('express-async-handler')
 const { findById } = require('../models/goalModel')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 // @desc    Get Goals
 // @route   Get /api/goals
 // @access  Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -20,7 +21,8 @@ const setGoal = asyncHandler(async(req, res) => {
         throw new Error('please add a text field')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -35,6 +37,18 @@ const updateGoal = asyncHandler(async(req, res) => {
         throw new error('goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+    //Maker sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true
     })
@@ -50,6 +64,18 @@ const deleteGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(400)
         throw new error('Goal not found')
+    }
+    const user = await User.findById(req.user.id)
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('user not found')
+    }
+    //Maker sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('User not authorized')
     }
     const deletedGoal = await Goal.findByIdAndDelete(req.params.id)
     res.status(200).json(deletedGoal)
